backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Fixture-backed FastAPI backend for AF Mol* Viewer"
authors = [
    { name = "Leo Kaindl", email = "leo.kaindl@helmholtz-munich.de" }
]
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.87",
    "fastapi>=0.115.0",
    "numpy>=2.4.4",
    "uvicorn>=0.30.0",
]

[project.scripts]
backend = "backend.api:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
